Issue #2930262 by t.murphy, penyaskito, shellshocked59: Create a view for all related entities of a given node so bulk operations can be performed on them

+++ src/LingotekContentTranslationService.php
@@ -785,7 +785,9 @@ class LingotekContentTranslationService implements LingotekContentTranslationSer
 
     $doc_id = $this->getDocumentId($entity);
     $metadata = $entity->lingotek_metadata->entity;
-    $metadata->delete();
+    if ($metadata !== NULL){
+      $metadata->delete();
+    }
   }
 
   /**
